# CBSSupport.API/OpenApi/cbs-support-api.baseline.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "CBS Support API",
    "description": "Versioned support and ticketing API. All operations require an authenticated Client or Admin principal unless stated otherwise. Client operations are tenant-scoped from trusted claims; Admin operations require the Admin policy and selected-tenant authorization where applicable.",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/notifications": {
      "get": {
        "tags": [
          "Notifications"
        ],
        "operationId": "get_api_v1_notifications",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Number of records to return (1-100; default 50 for conversations and 20 for notifications).",
            "schema": {
              "maximum": 100,
              "minimum": 1,
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Opaque keyset cursor returned by the previous page.",
            "schema": {
              "$ref": "#/components/schemas/PageCursor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationPage"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationPage"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationPage"
                }
              }
            }
          },
          "400": {
            "description": "Request validation failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "The authenticated principal is not authorized for this operation.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "The resource does not exist or is not visible to the caller.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected server error occurred.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notifications/{notificationId}/read": {
      "put": {
        "tags": [
          "Notifications"
        ],
        "operationId": "put_api_v1_notifications__notificationId__read",
        "parameters": [
          {
            "name": "notificationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationChangedEvent"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationChangedEvent"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationChangedEvent"
                }
              }
            }
          },
          "400": {
            "description": "Request validation failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "The authenticated principal is not authorized for this operation.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "The resource does not exist or is not visible to the caller.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected server error occurred.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "The request conflicts with current resource state or its concurrency token.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": [ ],
            "csrfToken": [ ]
          },
          {
            "bearerAuth": [ ]
          }
        ]
      }
    },
    "/api/v1/notifications/read-all": {
      "put": {
        "tags": [
          "Notifications"
        ],
        "operationId": "put_api_v1_notifications_read_all",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationBulkReadResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationBulkReadResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationBulkReadResult"
                }
              }
            }
          },
          "400": {
            "description": "Request validation failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "The authenticated principal is not authorized for this operation.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "The resource does not exist or is not visible to the caller.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected server error occurred.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "The request conflicts with current resource state or its concurrency token.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": [ ],
            "csrfToken": [ ]
          },
          {
            "bearerAuth": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string",
            "format": "uri-reference"
          },
          "traceId": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        },
        "description": "RFC 7807 problem response."
      },
...
      "PageCursor": {
        "maxLength": 1024,
        "type": "string",
        "description": "Opaque keyset cursor returned by the previous page. Do not inspect or modify it."
      },
...
      "NotificationBulkReadResult": {
        "required": [
          "count",
          "unreadCount"
        ],
        "type": "object",
        "properties": {
          "count": {
            "type": "integer",
            "format": "int64"
          },
          "unreadCount": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "NotificationChangedEvent": {
        "required": [
          "notification",
          "unreadCount"
        ],
        "type": "object",
        "properties": {
          "notification": {
            "$ref": "#/components/schemas/NotificationResponse2"
          },
          "unreadCount": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "NotificationPage": {
        "required": [
          "items",
          "nextCursor",
          "unreadCount"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NotificationResponse"
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true
          },
          "unreadCount": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "NotificationResponse": {
        "required": [
          "id",
          "caseId",
          "eventType",
          "createdAt",
          "readAt",
          "title",
          "message"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "caseId": {
            "type": "integer",
            "format": "int64"
          },
          "eventType": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "readAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "title": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "NotificationResponse2": {
        "required": [
          "id",
          "caseId",
          "eventType",
          "createdAt",
          "readAt",
          "title",
          "message"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "caseId": {
            "type": "integer",
            "format": "int64"
          },
          "eventType": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "readAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "title": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "nullable": true
      },
...
    },
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "description": "Browser session cookie. Unsafe browser mutations also require the RequestVerificationToken header.",
        "name": "CBSSupport.AuthCookie",
        "in": "cookie"
      },
      "bearerAuth": {
        "type": "http",
        "description": "Optional JWT bearer authentication for explicitly enabled external API clients.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "csrfToken": {
        "type": "apiKey",
        "description": "Required for unsafe cookie-authenticated browser requests.",
        "name": "RequestVerificationToken",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "cookieAuth": [ ]
    },
    {
      "bearerAuth": [ ]
    }
  ],
...
}